# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Documentation API",
        "version": "1.0"
    },
    "paths": {
        "/api/maps": {
            "post": {
                "tags": [
                    "Maps"
                ],
                "summary": "Maps",
                "description": "Masukkan Kecamatan / Kota / Provinsi / Kode Pos",
                "operationId": "Maps",
                "requestBody": {
                    "description": "",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "value"
                                ],
                                "properties": {
                                    "value": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "default": {
                        "description": ""
                    }
                }
            }
        },
        "/api/rates": {
            "post": {
                "tags": [
                    "Rates"
                ],
                "summary": "Cek Tarif Pengiriman",
                "description": "Masukkan Asal Postal Code, Tujuan Postal Code, Items",
                "operationId": "Rates",
                "requestBody": {
                    "description": "Form",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [],
                                "properties": {
                                    "asal_postal_code": {
                                        "type": "integer"
                                    },
                                    "tujuan_postal_code": {
                                        "type": "integer"
                                    },
                                    "items": {
                                        "type": "array",
                                        "items": {
                                            "properties": {
                                                "name": {
                                                    "type": "string"
                                                },
                                                "value": {
                                                    "type": "integer"
                                                },
                                                "quantity": {
                                                    "type": "integer"
                                                },
                                                "weight": {
                                                    "type": "integer"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "default": {
                        "description": ""
                    }
                }
            }
        },
...
        "/api/detail-kategori-toko": {
            "post": {
                "tags": [
                    "Toko"
                ],
                "summary": "Kategori",
                "description": "Menampilkan semua produk berdasarkan kategori toko yg dipillih",
                "operationId": "Kategori",
                "requestBody": {
                    "description": "",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "kategori_id",
                                    "toko_id"
                                ],
                                "properties": {
                                    "kategori_id": {
                                        "type": "integer"
                                    },
                                    "toko_id": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "default": {
                        "description": "return array model produk"
                    }
                }
            }
        },
...
        "/api/cost": {
            "post": {
                "tags": [
                    "Tarif Pengiriman"
                ],
                "summary": "tarif pengiriman",
                "description": "masukkan id_kota_asal, id_kota_tujuan, total_berat_produk, kurir",
                "operationId": "tarif pengiriman",
                "requestBody": {
                    "description": "",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "id_kota_asal": {
                                        "type": "integer"
                                    },
                                    "id_kota_tujuan": {
                                        "type": "integer"
                                    },
                                    "total_berat_produk": {
                                        "type": "integer"
                                    },
                                    "kurir": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "default": {
                        "description": ""
                    }
                }
            }
        }
    }
}